+++ test/stx_multi_select_field_test.dart
@@ -36,6 +36,7 @@ void main() {
         options: invalidOptions,
         forceValueToSet: true);
 
+    expect(field.state.initialValue, initialValue);
     expect(field.state.value, initialValue);
     expect(field.state.options, invalidOptions);
   });
@@ -78,6 +79,7 @@ void main() {
 
     field.changeValue(invalidValue);
 
+    expect(field.state.initialValue, initialValue);
     expect(field.state.value, invalidValue);
     expect(field.state.options, options);
   });